💥 Merge lock-directory into the filename, support formatting
Replace `lock-directory`/`lock-filename` with one `lock-filename` option, and also enable context formatting for it.

I think it is more convenient to specify the path to the config file once, rather than through two separate parameters.

Now one can write this, once:

    lock-filename = "requirements/{env_name}.lock"

And, because we enabled [context formatting] the `env_name` will be actually substituted to the current environment name. Which can then be [inherited as a template].

[context formatting]: https://hatch.pypa.io/1.7/config/environment/advanced/
[inherited as a template]: https://hatch.pypa.io/1.7/config/environment/overview/#inheritance